+++ Sources/VComponents/Presentation/Components/Buttons/Stretched Button (Loading)/VLoadingStretchedButton.swift
@@ -139,33 +139,37 @@ public struct VLoadingStretchedButton<CustomLabel>: View where CustomLabel: View
     private func labelView(
         internalState: VLoadingStretchedButtonInternalState
     ) -> some View {
-        HStack(spacing: appearance.labelAndSpinnerSpacing) {
+        HStack(spacing: 0) {
+            if !appearance.labelAndSpinnerSpacingType.hasFlexibleSpacing {
+                Spacer()
+            }
+            
             if internalState == .loading {
                 switch appearance.spinnerPlacement {
                 case .leading: spinnerView(internalState: internalState)
                 case .center: EmptyView()
-                case .trailing: spinnerCompensatorView(internalState: internalState)
+                case .trailing: spinnerCompensatorView
                 }
             }
+            
+            Spacer().frame(width: appearance.labelAndSpinnerSpacingType.spacing)
 
-            if
-                internalState == .loading,
-                appearance.spinnerPlacement == .center
-            {
-                ZStack {
+            ZStack {
+                if
+                    internalState == .loading,
+                    appearance.spinnerPlacement == .center
+                {
+                    
                     spinnerView(internalState: internalState)
-                }
-                .frame(maxWidth: .infinity)
-                
-            } else {
-                Group {
+                    
+                } else {
                     switch label {
                     case .title(let title):
                         labelTextElement(internalState: internalState, title: title)
-
+                        
                     case .image(let image):
                         labelImageElement(internalState: internalState, image: image)
-
+                        
                     case .titleAndImage(let title, let image):
                         switch appearance.labelTextAndLabelImagePlacement {
                         case .textAndImage:
@@ -173,27 +177,33 @@ public struct VLoadingStretchedButton<CustomLabel>: View where CustomLabel: View
                                 labelTextElement(internalState: internalState, title: title)
                                 labelImageElement(internalState: internalState, image: image)
                             }
-
+                            
                         case .imageAndText:
                             HStack(spacing: appearance.labelSpacing) {
                                 labelImageElement(internalState: internalState, image: image)
                                 labelTextElement(internalState: internalState, title: title)
                             }
                         }
-
+                        
                     case .custom(let builder):
                         builder(internalState)
                     }
                 }
-                .frame(maxWidth: .infinity)
             }
+            .frame(maxWidth: appearance.labelAndSpinnerSpacingType.hasFlexibleSpacing ? .infinity : nil)
+            
+            Spacer().frame(width: appearance.labelAndSpinnerSpacingType.spacing)
             
             if internalState == .loading {
                 switch appearance.spinnerPlacement {
-                case .leading: spinnerCompensatorView(internalState: internalState)
+                case .leading: spinnerCompensatorView
                 case .center: EmptyView()
                 case .trailing: spinnerView(internalState: internalState)
                 }
+            }
+            
+            if !appearance.labelAndSpinnerSpacingType.hasFlexibleSpacing {
+                Spacer()
             }
         }
         .scaleEffect(internalState == .pressed ? appearance.labelPressedScale : 1)
@@ -222,9 +232,7 @@ public struct VLoadingStretchedButton<CustomLabel>: View where CustomLabel: View
         VContinuousSpinner(appearance: appearance.spinnerAppearance)
     }
     
-    private func spinnerCompensatorView(
-        internalState: VLoadingStretchedButtonInternalState
-    ) -> some View {
+    private var spinnerCompensatorView: some View {
         Spacer()
             .frame(width: appearance.spinnerAppearance.dimension)
     }
@@ -318,13 +326,15 @@ public struct VLoadingStretchedButton<CustomLabel>: View where CustomLabel: View
     }
 }
 
-#Preview("Spinner Placements") {
+#Preview("Spinner Configurations") {
+    @Previewable @State var labelAndSpinnerSpacingType: VLoadingStretchedButtonAppearance.LabelAndSpinnerSpacingType = .stretched(spacing: 20)
     @Previewable @State var placement: VLoadingStretchedButtonAppearance.SpinnerPlacement = .leading
 
     PreviewContainer {
         VLoadingStretchedButton(
             appearance: {
                 var appearance: VLoadingStretchedButtonAppearance = .init()
+                appearance.labelAndSpinnerSpacingType = labelAndSpinnerSpacingType
                 appearance.spinnerPlacement = placement
                 return appearance
             }(),
@@ -343,6 +353,16 @@ public struct VLoadingStretchedButton<CustomLabel>: View where CustomLabel: View
                             return
                         }
                         
+                        labelAndSpinnerSpacingType = .stretched(spacing: 20)
+                        placement = .leading
+                        
+                        do {
+                            try await Task.sleep(for: .seconds(1))
+                        } catch {
+                            return
+                        }
+                        
+                        labelAndSpinnerSpacingType = .fixed(spacing: 20)
                         placement = .leading
 
                         do {
@@ -359,6 +379,16 @@ public struct VLoadingStretchedButton<CustomLabel>: View where CustomLabel: View
                             return
                         }
                         
+                        labelAndSpinnerSpacingType = .fixed(spacing: 20)
+                        placement = .trailing
+                        
+                        do {
+                            try await Task.sleep(for: .seconds(1))
+                        } catch {
+                            return
+                        }
+                        
+                        labelAndSpinnerSpacingType = .stretched(spacing: 20)
                         placement = .trailing
                     }
                 }